Separate physical model forr DB storage from the logical model.
Introduce io-ts Decoders to decode data from the sequelize objects read from the database. This has the effect of cleaning up the objects before being returned from the model, but will detect any issues in translating data from the database.

This approach was taken as it is likelly we will want to store structured data, stored as a JSON string rather than relations between tables, as part of an Event Motion. The physical model will be able to store the JSON string, but the logical model will have the expanded structure.